refactor(site/system): simplify the file structure
