+++ site/src/system/styles/index.ts
@@ -1,0 +1,2 @@
+export * from './reset.css';
+export * from './sprinkles.css';
